[Profiler] Build call tree in `collection.cpp`
Pull Request resolved: https://github.com/pytorch/pytorch/pull/77698

This PR adds tree building to the post processing of profiler. The basic algorithm is to sort the events, maintain a stack and a priority queue of event ends, and push/pop accordingly. The logic for merging Python events is still separate in `profiler_kineto.cpp`. That can be removed when Python events have an `EventType`.
ghstack-source-id: 156788606

Differential Revision: [D36321105](https://our.internmc.facebook.com/intern/diff/D36321105/)